Added props to customize messaging inside of permission buttons.
Users have been confused in some of our apps — when they see the button that says ‘DENIED X’ it wasn’t intuitive to them that they could tap that button to resolve the problem. I added some public vars that provide formats defaulting to the existing messaging. I believe this might be useful to others who wish to customize the text in the permission view further as we needed.